feat(ton-provider): init implemented get_block_by_height

+++ rust/main/chains/hyperlane-ton/src/traits/ton_api_center.rs
@@ -80,7 +80,7 @@ pub trait TonApiCenter {
         &self,
         workchain: i32,
         shard: Option<String>,
-        seqno: Option<u32>,
+        seqno: Option<u64>,
         mc_seqno: Option<u32>,
         start_utime: Option<i64>,
         end_utime: Option<i64>,
